finds and prints the largest prime factor of the number 612852475143

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -11,7 +11,7 @@
 
 int main(void)
 {
-	long long index, index2, number, max;
+	unsigned long index, index2, number, max;
 	int is_prime;
 
 	number = 612852475143;
